+++ internal/database/exams.sql.go
@@ -29,6 +29,7 @@ WITH upsert as (
     VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
     ON CONFLICT (site_id, accession) DO UPDATE
     SET
+        updated_at = CURRENT_TIMESTAMP,
         visit_id = EXCLUDED.visit_id,
         mrn_id = EXCLUDED.mrn_id,
         procedure_id = EXCLUDED.procedure_id,
@@ -306,6 +307,26 @@ func (q *Queries) GetExamBySiteIDAccession(ctx context.Context, arg GetExamBySit
 		&i.SiteIsCms,
 	)
 	return i, err
+}
+
+const getExamIDBySiteIDAccession = `-- name: GetExamIDBySiteIDAccession :one
+SELECT id
+FROM exams
+WHERE
+    site_id = $1
+    AND accession = $2
+`
+
+type GetExamIDBySiteIDAccessionParams struct {
+	SiteID    pgtype.Int4
+	Accession string
+}
+
+func (q *Queries) GetExamIDBySiteIDAccession(ctx context.Context, arg GetExamIDBySiteIDAccessionParams) (int64, error) {
+	row := q.db.QueryRow(ctx, getExamIDBySiteIDAccession, arg.SiteID, arg.Accession)
+	var id int64
+	err := row.Scan(&id)
+	return id, err
 }
 
 const updateExam = `-- name: UpdateExam :one