+++ sql/schema/018_exams_new_idx.sql
@@ -1,0 +1,5 @@
+-- +goose Up
+CREATE UNIQUE INDEX IF NOT EXISTS exams_site_id_accession ON exams(site_id, accession);
+
+-- +goose Down
+DROP INDEX IF EXISTS exams_site_id_accession;